Move shared package and fs helpers into node/utils.
